Change LinkContainer to Link if child is a Anchor
LinkContainer is meant for React-Bootstrap components to
be used as navigation widgets with React-Router. But
using it with Anchors (<a>) will cause an error since
the prop active will be add to the component.
Instead of using LinkContainer with Anchor the
React-Router component Link should be used.

Also: I found a MenuItem not using LinkContainer
and fixed that.